+++ src/nnue/nnue_accumulator.h
@@ -37,10 +37,6 @@ class Position;
 
 namespace Stockfish::Eval::NNUE {
 
-using BiasType       = std::int16_t;
-using PSQTWeightType = std::int32_t;
-using IndexType      = std::uint32_t;
-
 template<IndexType Size>
 struct alignas(CacheLineSize) Accumulator;
